Migrate documentation to GitHub
Easier to maintain the documentation if it is stored in the
GitHub README file.  Next plugin release after this is merged
will include the documentation in the official plugins page
https://plugins.jenkins.io/openedge/ and will remove the warning that
says:

> This content is served from the Jenkins Wiki Export which is now
> permanently offline and before that a read-only state. We would love
> your help in moving plugin documentation to GitHub, see the guidelines.

Will improve the plugin health score as reported at
https://plugins.jenkins.io/openedge/healthscore/

https://www.jenkins.io/doc/developer/tutorial-improve/migrate-documentation-to-github/
is the tutorial that was used to create this pull request.

Extracted the changelog for the most recent releases by reading the git history.

Needs:

* https://github.com/jenkinsci/openedge-plugin/pull/7